+++ flow-typed/lib/report-utils.js
@@ -1,0 +1,9 @@
+declare type ReportColumn = {
+  accessor?: string | (Object => string),
+  exportHeaderName?: string,
+};
+
+declare type ReportOptions = {
+  delimiter?: string,
+  fileMimeType?: string,
+};