+++ src/index.js
@@ -132,3 +132,5 @@ export {
   updateSearchUrl,
 } from "./utils/url-utils";
 export { Warning } from "./utils/warning-utils";
+
+export { toDelimitedReport, toTableRow } from "./utils/table-utils";